refactor(payment service): migrate from PostgreSQL to MongoDB

+++ payment-service/src/main/java/paymentservice/dto/PaymentResponse.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import paymentservice.entity.PaymentStatus;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record PaymentResponse(
@@ -20,5 +21,5 @@ public record PaymentResponse(
 
         Instant timestamp,
 
-        Long paymentAmount) {
+        BigDecimal paymentAmount) {
 }
